Create string class and remove unnecesary casts

+++ ARPB2Engine/Model/Number.cs
@@ -11,7 +11,7 @@ namespace ARPB2Engine.Model
 
         public bool Equals(Number other)
         {
-            return (int)value == (int)other.value;
+            return value == other.value;
         }
